DR-333 Code changes for Dismissed Alert

+++ src/main/java/com/wheelsup/alertservice/domain/PhoneContact.java
@@ -14,8 +14,8 @@ public class PhoneContact {
     @Id
     @GeneratedValue
     @Column(name = "phoneContactId")
-    private Integer id;
+    private Long id;
     private String phoneNumber;
-    private Integer memberId;
+    private Long memberId;
 
 }
